pyproject.toml
[project]
name = "astroeasy"
authors = [
  { name="Zach", email="jzg@alum.mit.edu" },
]
version = "0.2.0"
description = "Astrometry.net made easy - plate solving and index management"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE*"]
keywords = ["astronomy", "astrometry", "plate-solving", "wcs"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Astronomy",
]

dependencies = [
    "astropy>=5.0",
    "numpy>=1.20",
    "pyyaml>=6.0",
    "tqdm>=4.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.4.0",
    "twine>=6.2.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "genbadge[coverage,tests]>=1.1.3",
]

[project.scripts]
astroeasy = "astroeasy.cli:main"

[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/zgazak/astroeasy"
Issues = "https://github.com/zgazak/astroeasy/issues"
Documentation = "https://github.com/zgazak/astroeasy#readme"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["astroeasy"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["astroeasy"]
